+++ mediation/MyTarget/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## myTarget Unity Mediation Plugin Changelog
 
+#### Next Version
+- Moved adapter content to `GoogleMobileAds/Mediation/MyTarget/`.
+- Refactored adapter namespace to use `GoogleMobileAds.Mediation.MyTarget`.
+
 #### [Version 3.13.0](https://dl.google.com/googleadmobadssdk/mediation/unity/mytarget/MyTargetUnityAdapter-3.13.0.zip)
 - Supports [myTarget Android adapter version 5.16.4.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/mytarget/CHANGELOG.md#version-51640).
 - Supports [myTarget iOS adapter version 5.17.4.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/MyTarget/CHANGELOG.md#version-51740).